@@ -1,16 +1,16 @@
 module github.com/sv-tools/conf-reader-flags
 
-go 1.23.0
+go 1.24.0
 
 require (
 	github.com/spf13/pflag v1.0.6
 	github.com/stretchr/testify v1.10.0
-	github.com/sv-tools/conf v1.3.0
+	github.com/sv-tools/conf v1.4.0
 )
 
 require (
 	github.com/davecgh/go-spew v1.1.1 // indirect
 	github.com/pmezard/go-difflib v1.0.0 // indirect
-	github.com/spf13/cast v1.7.0 // indirect
+	github.com/spf13/cast v1.7.1 // indirect
 	gopkg.in/yaml.v3 v3.0.1 // indirect
 )
